Adds Gates and Policy classes for blogpost so users can only edit and update if they have access to them

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\BlogPost;
 use App\Http\Requests\StorePost;
 //use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -47,6 +48,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        //$this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -90,6 +92,11 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
+        // if(Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post!");
+        // }
+
+        $this->authorize($post);
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -103,6 +110,12 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if(Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post!");
+        // }
+        $this->authorize($post);
+
         $validatedData = $request->validated();
 
         $post->fill($validatedData);
@@ -121,6 +134,10 @@ class PostController extends Controller
     public function destroy(Request $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+        // if(Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog post!");
+        // }
+        $this->authorize($post);
         $post->delete();
         $request->session()->flash('status', 'Blog post was deleted!');
